feat(tenpayv3): 新增直连商户委托代扣相关接口

+++ src/SKIT.FlurlHttpClient.Wechat.TenpayV3/ExtendedSDK/Global/Events/PAPay/PAPayContractResource.cs
@@ -41,7 +41,7 @@ namespace SKIT.FlurlHttpClient.Wechat.TenpayV3.ExtendedSDK.Global.Events
         /// </summary>
         [Newtonsoft.Json.JsonProperty("plan_id")]
         [System.Text.Json.Serialization.JsonPropertyName("plan_id")]
-        public int PlanId { get; set; }
+        public long PlanId { get; set; }
 
         /// <summary>
         /// 获取或设置用户唯一标识。
