+++ src/SKIT.FlurlHttpClient.Wechat.TenpayV3/ExtendedSDK/Global/Models/PAPay/PresignPAPayContractEntrustAppRequest.cs
@@ -19,7 +19,7 @@ namespace SKIT.FlurlHttpClient.Wechat.TenpayV3.ExtendedSDK.Global.Models
         /// </summary>
         [Newtonsoft.Json.JsonProperty("plan_id")]
         [System.Text.Json.Serialization.JsonPropertyName("plan_id")]
-        public int PlanId { get; set; }
+        public long PlanId { get; set; }
 
         /// <summary>
         /// 获取或设置商户侧签约协议号。
