+++ src/SKIT.FlurlHttpClient.Wechat.TenpayV3/Models/_Partner/Ecommerce/CombinePAPay/GetEcommerceCombinePAPayContractByOutContractCodeRequest.cs
@@ -10,7 +10,7 @@ namespace SKIT.FlurlHttpClient.Wechat.TenpayV3.Models
         /// </summary>
         [Newtonsoft.Json.JsonIgnore]
         [System.Text.Json.Serialization.JsonIgnore]
-        public int PlanId { get; set; }
+        public long PlanId { get; set; }
 
         /// <summary>
         /// 获取或设置商户侧签约协议号。
