+++ src/SKIT.FlurlHttpClient.Wechat.TenpayV3/Models/_Partner/Ecommerce/CombinePAPay/PresignEcommerceCombinePAPayContractEntrustAppRequest.cs
@@ -17,7 +17,7 @@ namespace SKIT.FlurlHttpClient.Wechat.TenpayV3.Models
         /// </summary>
         [Newtonsoft.Json.JsonProperty("plan_id")]
         [System.Text.Json.Serialization.JsonPropertyName("plan_id")]
-        public int PlanId { get; set; }
+        public long PlanId { get; set; }
 
         /// <summary>
         /// 获取或设置商户签约协议号。
